Bug fixed: supercell

+++ src/pypolymlp/utils/supercell_utils.py
@@ -40,7 +40,7 @@ def get_supercell(
                  If the supercell matrix is non-diagonal,
                  phonopy algorithm is automatically used.
     """
-    size = np.array(supercell_matrix)
+    size = np.array(supercell_matrix).astype(int)
     if size.shape == (3, 3):
         if not _is_diagonal(size):
             from pypolymlp.utils.phonopy_utils import phonopy_supercell
@@ -80,7 +80,7 @@ def _get_supercell_diagonal(
 
     # Loop sequence: z -> y -> x, which is compatible with phonopy
     nx, ny, nz = size
-    trans_all = np.indices((int(nz), int(ny), int(nx))).reshape(3, -1).T
+    trans_all = np.indices((nz, ny, nx)).reshape(3, -1).T
     trans_all = trans_all[:, [2, 1, 0]]
     positions_new = (st.positions.T[:, None] + trans_all[None, :]).reshape((-1, 3))
     sup.positions = (positions_new / size).T
